Fix typos(?) found by codespell
These typos could have been generated by some program, in which case
the program should be fixed instead. Yet the typos are inconcistent,
as if the output had been manually copied. Why is the output poorly
aligned?

Use seconds or s for seconds, consistently.

Use ops for operations, consistently. Unless you want the singular
op for operation, as the value is often < 2.